requirements.txt
streamlit
google-generativeai
Pillow
scikit-learn
